scripts/ingestion/requirements.txt
py2neo
tqdm
python-dotenv
rdflib
boto3
requests